+++ frontend/src/component/application/Application.tsx
@@ -1,5 +1,5 @@
 /* eslint react/no-multi-comp:off */
-import React, { lazy, useContext, useState } from 'react';
+import React, { useContext, useState } from 'react';
 import {
     Box,
     Avatar,
@@ -31,6 +31,7 @@ import { formatUnknownError } from 'utils/formatUnknownError';
 import { useRequiredPathParam } from 'hooks/useRequiredPathParam';
 import { useUiFlag } from 'hooks/useUiFlag';
 import { ApplicationEdit } from './ApplicationEdit/ApplicationEdit';
+import ApplicationOverview from './ApplicationOverview';
 
 type Tab = {
     title: string;
@@ -67,8 +68,6 @@ const StyledTab = styled(Tab)(({ theme }) => ({
         minWidth: 160,
     },
 }));
-
-const ApplicationOverview = lazy(() => import('./ApplicationOverview'));
 
 export const Application = () => {
     const useOldApplicationScreen = !useUiFlag('sdkReporting');
